Refactor script to use typescript, put doc gen output into a temporary folder and upload resulting json to S3
