+++ .doxityrc
@@ -1,7 +1,7 @@
 {
-  "target": "scripts/doxity",
+  "target": "temporary/doxity",
   "src": "contracts/*",
   "dir": "pages/docs",
   "source": "https://github.com/DigixGlobal/doxity-gatsby-starter-project/archive/a4886b7a7a04c018ac04fed3125d7d4785e74bed.tar.gz",
-  "out": "scripts/doxity/docs"
+  "out": "temporary/doxity/docs"
 }
