+++ tsconfig.json
@@ -20,6 +20,7 @@
         "./globals.d.ts",
         "./util/**/*",
         "./test/**/*",
-        "./migrations/**/*"
+        "./migrations/**/*",
+        "./scripts/**/*"
     ]
 }
